slå sammen producer og initial load-task for tiltakstyper
det er uansett bare fra denne tasken at meldinger blir produsert til
siste-tiltakstyper topic
